start moving configs. work on Base class
