--- Procfile
@@ -1,1 +1,0 @@
-web: gunicorn tinker:app --log-file=-
