+++ db_migrate.py
@@ -1,9 +1,12 @@
 #!/Users/ejc84332/envs/tinker/bin/python
 import imp
+
 from migrate.versioning import api
+
+from config.config import SQLALCHEMY_DATABASE_URI
+from config.config import SQLALCHEMY_MIGRATE_REPO
 from tinker import db
-from config import SQLALCHEMY_DATABASE_URI
-from config import SQLALCHEMY_MIGRATE_REPO
+
 migration = SQLALCHEMY_MIGRATE_REPO + '/versions/%03d_migration.py' % (api.db_version(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO) + 1)
 tmp_module = imp.new_module('old_model')
 old_model = api.create_model(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO)
