--- runtime.txt
@@ -1,1 +1,0 @@
-python-2.7.9
